+++ src/components/Navbar.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import logo from '../logo.svg';
 
 const Navbar = () => {
     return (
